Create a20181026_killMega.ino
Because the inputs from the mainMega are set as INPUT_PULLUP, these pins will naturally be pulled up (HIGH) in the situation that the mainMega is floating/killed/disconnected.  Because this is the case, you want your code to execute the "safe" behaviors when these pins are high; i.e. you want the system to KILL "remoteKillInPin" reads HIGH, and UNKILL when the pin is pulled LOW (because it takes "effort" to pull the pin LOW from it's naturally HIGH state).  I changed the code to reflect this behavior.

I also changed the system so that the stateInPin reads HIGH in manual, and LOW for autonomous (since we want the "natural" state to be in manual override).  The lights were changed to reflect this.

Also increasing the delay from 50 to 100 helped with the flickering problem (since you ideally want this killMega to loop slightly slower than the mainMega).