2019.01.31 version of simulation and estimator
Robot simulation is fairly mature, and based off of similar past versions.  The estimator still needs work, particularly in the state-transition ("A") matrix.  The current version has two different state-transition matrices.  The first is a simple constant-acceleration state transition matrix, the second is a state transition matrix that predicts the two linear velocity states based on the position state predictions.