Create navprac.py
The concept I was going for with control loops was to run two in unison for the angle error and distance error while outside some radius of the waypoint. Then, within that waypoint, ignore the angle error and just run one control loop for distance, so that the vehicle would for forward and back to stay near the way point. If the vehicle drifts outside of the defined radius of waypoint, it would then restart the dual control loop to turn the boat and get back inside the inner radius. This (hopefully) would reduce the occurrence of bulky, wide turns right near the waypoint.